feat: CLIN-1207 - refractoring
