+++ src/main/java/bio/ferlab/clin/portal/forms/models/builders/AnalysisBuilder.java
@@ -1,27 +1,28 @@
-package bio.ferlab.clin.portal.forms.utils;
+package bio.ferlab.clin.portal.forms.models.builders;
 
 import bio.ferlab.clin.portal.forms.mappers.SubmitToFhirMapper;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
+import org.hl7.fhir.r4.model.ClinicalImpression;
 import org.hl7.fhir.r4.model.Patient;
 import org.hl7.fhir.r4.model.ServiceRequest;
 
 @AllArgsConstructor
-public class SequencingBuilder {
+public class AnalysisBuilder {
   
   private final SubmitToFhirMapper mapper;
   private final String panelCode;
   private final Patient patient;
-  private final ServiceRequest analysis;
-  
+  private ClinicalImpression clinicalImpression;
+
   public Result build() {
-    final ServiceRequest serviceRequest = mapper.mapToSequencing(panelCode, patient, analysis);
+    final ServiceRequest serviceRequest = mapper.mapToAnalysis(panelCode, patient, clinicalImpression);
     return new Result(serviceRequest);
   }
-  
+
   @AllArgsConstructor
   @Getter
   public static class Result {
-    private final ServiceRequest sequencing;
+    private final ServiceRequest analysis;
   }
 }
