+++ src/main/java/bio/ferlab/clin/portal/forms/models/builders/ClinicalImpressionBuilder.java
@@ -1,28 +1,25 @@
-package bio.ferlab.clin.portal.forms.utils;
+package bio.ferlab.clin.portal.forms.models.builders;
 
 import bio.ferlab.clin.portal.forms.mappers.SubmitToFhirMapper;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import org.hl7.fhir.r4.model.ClinicalImpression;
 import org.hl7.fhir.r4.model.Patient;
-import org.hl7.fhir.r4.model.ServiceRequest;
 
 @AllArgsConstructor
-public class AnalysisBuilder {
+public class ClinicalImpressionBuilder {
   
   private final SubmitToFhirMapper mapper;
-  private final String panelCode;
   private final Patient patient;
-  private ClinicalImpression clinicalImpression;
-
+  
   public Result build() {
-    final ServiceRequest serviceRequest = mapper.mapToAnalysis(panelCode, patient, clinicalImpression);
-    return new Result(serviceRequest);
+    final ClinicalImpression clinicalImpression = mapper.mapToClinicalImpression(patient);
+    return new Result(clinicalImpression);
   }
-
+  
   @AllArgsConstructor
   @Getter
   public static class Result {
-    private final ServiceRequest analysis;
+    private final ClinicalImpression clinicalImpression;
   }
 }
